backend/requirements.txt
flask 
flask-sqlalchemy 
SQLAlchemy
flask-cors 
python-dotenv 
psycopg2-binary
lask-JWT-Extended 